backend/requirements.txt
fastapi==0.135.1
uvicorn[standard]==0.42.0
SQLAlchemy==2.0.48
psycopg2-binary==2.9.11
python-dotenv==1.2.2
pydantic==2.12.5
email-validator==2.3.0
python-multipart==0.0.22
fpdf2==2.8.2